Add files via upload
Attached is the zip file containing my PCB design for converting AC to DC. It has been designed fully on KiCad 8.0. 
This is my first PCB design. I have designed it to fit a 150 X 100 mm copper clad for etching. 

The purpose of this design is to convert 12 V AC to 5V DC as well as 12V DC. 
Kindly Note: The stepdown transformer to convert 230V AC to 12 V AC hasn't been included here in this design but there is sufficient space available on the board to accommodate for it's installation.